### 2022-09-01 #4732 @Kevin-Mattheus-Moerman
@tmigot this is just a small comment after having a look at your README, it says:
`PDENLPModels specializes the [NLPModel API](https://github.com/JuliaSmoothOptimizers/NLPModels.jl) to optimization problem with partial differential equation in the constraints.`
Check if I am right, but it seems like perhaps it should read:
`PDENLPModels specializes the [NLPModel API](https://github.com/JuliaSmoothOptimizers/NLPModels.jl) to optimization problems with partial differential equations in the constraints.`, i.e. it may need to use the plural form twice there.
