### 2022-09-01 #4732 @jbytecode
@Kevin-Mattheus-Moerman - Yet another beautiful Julia package comes again, and It seems this submission is not the first one of @tmigot, here is the [previous one](https://joss.theoj.org/papers/10.21105/joss.03991.pdf) which has been edited by me. Of course I am accepting being the handling editor, nowadays, everybody is in their summer holiday and I can take care with it with my cup of coffee :)
