+++ paper/paper.bib
@@ -327,7 +327,6 @@ doi={10.1137/16M1081063}
 
 @software{Krysl_FinEtools_jl_2021,
 author = {Krysl, Petr},
-doi = {10.5281/zenodo.1234},
 month = {12},
 title = {{FinEtools.jl}},
 url = {https://github.com/PetrKryslUCSD/FinEtools.jl},
@@ -381,6 +380,7 @@ year = {2021}
   pages={251--266},
   year={2012},
   publisher={De Gruyter},
+  doi = {10.1515/jnum-2012-0013},
   url = {https://wiki2.org/en/FreeFem%2B%2B}
 }
 
@@ -393,6 +393,7 @@ year = {2021}
   pages={24--es},
   year={2007},
   publisher={ACM New York, NY, USA},
+  doi = {10.1145/1268776.1268779},
   url = {https://dealii.org/}
 }
 
@@ -444,7 +445,7 @@ year = {2021}
       volume = {156},
       year = {2022},
       issn = {0098-1354},
-      doi = {https://doi.org/10.1016/j.compchemeng.2021.107567},
+      doi = {10.1016/j.compchemeng.2021.107567},
       url = {https://www.sciencedirect.com/science/article/pii/S0098135421003458},
       author = {Joshua L. Pulsipher and Weiqi Zhang and Tyler J. Hongisto and Victor M. Zavala},
       url = {https://github.com/pulsipher/InfiniteOpt.jl}
