+++ src/PDENLPModels.jl
@@ -91,21 +91,34 @@ import NLPModels:
   hess,
   cons,
   cons!,
+  cons_nln,
+  cons_nln!,
   jac,
+  jac_nln,
   jprod,
   jprod!,
+  jprod_nln,
+  jprod_nln!,
   jtprod,
   jtprod!,
+  jtprod_nln,
+  jtprod_nln!,
   jac_op,
   jac_op!,
+  jac_nln_op,
+  jac_nln_op!,
   hprod,
   hprod!,
   hess_op,
   hess_op!,
   jac_structure,
   jac_structure!,
+  jac_nln_structure,
+  jac_nln_structure!,
   jac_coord,
   jac_coord!,
+  jac_nln_coord,
+  jac_nln_coord!,
   hess_structure,
   hess_structure!,
   hess_coord,
@@ -120,21 +133,34 @@ export GridapPDENLPModel,
   hess,
   cons,
   cons!,
+  cons_nln,
+  cons_nln!,
   jac,
+  jac_nln,
   jprod,
   jprod!,
+  jprod_nln,
+  jprod_nln!,
   jtprod,
   jtprod!,
+  jtprod_nln,
+  jtprod_nln!,
   jac_op,
   jac_op!,
+  jac_nln_op,
+  jac_nln_op!,
   hprod,
   hprod!,
   hess_op,
   hess_op!,
   jac_structure,
   jac_structure!,
+  jac_nln_structure,
+  jac_nln_structure!,
   jac_coord,
   jac_coord!,
+  jac_nln_coord,
+  jac_nln_coord!,
   hess_structure,
   hess_structure!,
   hess_coord,
