+++ Project.toml
@@ -15,7 +15,7 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 FastClosures = "0.2, 0.3"
 ForwardDiff = "0.9.0, 0.10.0"
 Gridap = "0.15.5"
-NLPModels = "0.15, 0.16, 0.17, 0.18, 0.19"
+NLPModels = "0.19"
 julia = "^1.6.0"
 
 [extras]
