+++ previews/PR76/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"poisson-boltzman/#PDE-contrained-optimization","page":"PDE-constrained optimization","title":"PDE-contrained optimization","text":"","category":"section"},{"location":"poisson-boltzman/#Poisson-Boltzman-problem","page":"PDE-constrained optimization","title":"Poisson-Boltzman problem","text":"","category":"section"},{"location":"poisson-boltzman/","page":"PDE-constrained optimization","title":"PDE-constrained optimization","text":"We consider here a 2D Poisson-Boltzmann problem","category":"page"},{"location":"poisson-boltzman/","page":"PDE-constrained optimization","title":"PDE-constrained optimization","text":"leftlbrace\nbeginaligned\nminlimits_y in H_0^1(Omega) u in L^2(Omega)     frac12int_Omega y - y_d(x)^2 + frac12alpha int_Omega u^2dx \ntext st   -Delta y + sinhy = h + u quad textin  Omega=(-11)^2\n                   y = 0 quad textin  partialOmega\nendaligned\nright","category":"page"},{"location":"poisson-boltzman/","page":"PDE-constrained optimization","title":"PDE-constrained optimization","text":"with the forcing term h(x_1x_2)=-sin(omega x_1) sin(omega x_2), omega = pi - frac18, and target state","category":"page"},{"location":"poisson-boltzman/","page":"PDE-constrained optimization","title":"PDE-constrained optimization","text":"beginaligned\n    y_d(x) = begincases\n    10 quad textif  x in 025075^2 \n    5 quad textotherwise\n    endcases\nendaligned","category":"page"},{"location":"poisson-boltzman/","page":"PDE-constrained optimization","title":"PDE-constrained optimization","text":"The implementation as a GridapPDENLPModel is given as follows.","category":"page"},{"location":"poisson-boltzman/","page":"PDE-constrained optimization","title":"PDE-constrained optimization","text":"    using Gridap, PDENLPModels\n    #Domain\n    n = 100\n    model = CartesianDiscreteModel((-1,1,-1,1), (n,n))\n\n    #Definition of the spaces:\n    order = 1\n    valuetype = Float64\n    reffe = ReferenceFE(lagrangian, valuetype, order)\n    Xpde = TestFESpace(\n      model,\n      reffe;\n      conformity = :H1,\n      dirichlet_tags=\"boundary\",\n    )\n    Ypde = TrialFESpace(Xpde, 0.0)\n    Xcon = TestFESpace(model, reffe; conformity = :L2)\n    Ycon = TrialFESpace(Xcon)\n    Y = MultiFieldFESpace([Ypde, Ycon])\n\n    #Integration machinery\n    trian = Triangulation(model)\n    degree = 1\n    dΩ = Measure(trian,degree)\n\n    #Objective function:\n    yd(x) = min(x[1]-0.25, 0.75-x[1],x[2]-0.25, 0.75-x[2])>=0. ? 10. : 5.\n    function f(y, u)\n        ∫( 0.5 * (yd - y) * (yd - y) + 0.5 * 1e-4 * u * u )dΩ\n    end\n\n    #Definition of the constraint operator\n    ω = π - 1/8\n    h(x) = - sin(ω*x[1])*sin(ω*x[2])\n    function res(y, u, v)\n     ∫( ∇(v) ⊙ ∇(y) + (sinh ∘ y)*v - u*v - v * h )dΩ\n    end\n    op = FEOperator(res, Y, Xpde)\n    xin = zeros(Gridap.FESpaces.num_free_dofs(Y))\n    nlp = GridapPDENLPModel(xin, f, trian, Ypde, Ycon, Xpde, 
[…]
