### 2022-09-03 #4732 @HaoZeke
> 👋👋👋 Dear @HaoZeke and @matbesancon 👋👋👋
>
> Would you be willing to assist in reviewing this submission for **JOSS** (**Journal of Open Source Software**)?
>
> JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**PDENLPModels.jl: An NLPModel API for 1 Optimization Problems with PDE-Constraints**". You can find more information at the top of this Github issue (#4732).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. If you have any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a seperate thread. In that thread, there will be 20 check items for each single reviewer.
>
> Thank you in advance!

I'm willing to review as well :)
