### 2022-09-03 #4736 @editorialbot
[REVIEW]: PDENLPModels.jl: An NLPModel API for Optimization Problems with PDE-Constraints
**Submitting author:** @tmigot (<a href="http://orcid.org/0000-0001-7729-2513">Tangi Migot</a>)
**Repository:** https://github.com/JuliaSmoothOptimizers/PDENLPModels.jl
**Branch with paper.md** (empty if default branch): joss-paper-branch
**Version:** v0.3.3
**Archive:** https://doi.org/10.5281/zenodo.7447623
