### 2022-09-03 #4732 @jbytecode
Dear author @tmigot and reviewers @jmejia8, @HaoZeke

Please follow the review thread at https://github.com/openjournals/joss-reviews/issues/4736

Thank you in advance.
