+++ src/PDENLPModels.jl
@@ -91,20 +91,29 @@ import NLPModels:
   hess,
   cons,
   cons!,
+  cons_lin,
+  cons_lin!,
   cons_nln,
   cons_nln!,
   jac,
   jac_nln,
+  jac_lin,
   jprod,
   jprod!,
+  jprod_lin,
+  jprod_lin!,
   jprod_nln,
   jprod_nln!,
   jtprod,
   jtprod!,
+  jtprod_lin,
+  jtprod_lin!,
   jtprod_nln,
   jtprod_nln!,
   jac_op,
   jac_op!,
+  jac_lin_op,
+  jac_lin_op!,
   jac_nln_op,
   jac_nln_op!,
   hprod,
@@ -113,10 +122,14 @@ import NLPModels:
   hess_op!,
   jac_structure,
   jac_structure!,
+  jac_lin_structure,
+  jac_lin_structure!,
   jac_nln_structure,
   jac_nln_structure!,
   jac_coord,
   jac_coord!,
+  jac_lin_coord,
+  jac_lin_coord!,
   jac_nln_coord,
   jac_nln_coord!,
   hess_structure,
@@ -133,20 +146,29 @@ export GridapPDENLPModel,
   hess,
   cons,
   cons!,
+  cons_lin,
+  cons_lin!,
   cons_nln,
   cons_nln!,
   jac,
   jac_nln,
+  jac_lin,
   jprod,
   jprod!,
+  jprod_lin,
+  jprod_lin!,
   jprod_nln,
   jprod_nln!,
   jtprod,
   jtprod!,
+  jtprod_lin,
+  jtprod_lin!,
   jtprod_nln,
   jtprod_nln!,
   jac_op,
   jac_op!,
+  jac_lin_op,
+  jac_lin_op!,
   jac_nln_op,
   jac_nln_op!,
   hprod,
@@ -155,10 +177,14 @@ export GridapPDENLPModel,
   hess_op!,
   jac_structure,
   jac_structure!,
+  jac_lin_structure,
+  jac_lin_structure!,
   jac_nln_structure,
   jac_nln_structure!,
   jac_coord,
   jac_coord!,
+  jac_lin_coord,
+  jac_lin_coord!,
   jac_nln_coord,
   jac_nln_coord!,
   hess_structure,
