### 2022-09-24 08d3075
redefine `jac_op!` and `hess_op!`
