### 2022-09-24 e98d735
remove extra obj_weight
