### 2022-09-26 #4736 @jbytecode
Dear @jmejia8 and @HaoZeke,

I am so sorry if I am bothering you sending this post, it seems you have some progress checking some items off in your task lists, but if it is possible, could you please update your status and inform us on how is your review going?

Do you have any suggestions on both the software and paper?

Thank you in advance and have a nice week.
