### 2022-09-26 #4736 @jmejia8
Dear @tmigot and @jbytecode

I have finished reviewing the paper and code related to PDENLPModels. The structure of the document is clear, and the writing is acceptable. The package seems to be well-coded, tested, and organized. The installation, tests, and examples worked perfectly in my Julia distribution.

Below are some initial comments:

1. The documentation can be improved in presenting examples (I will open issues related to this in the original repo).
2. I did not find the **community guidelines**. Please add a section in the README (or separate) file covering this part.
3. I understand that the package has been designed to handle specific optimization problems. However, in the paper, it could be better to explain the problem before stating the mathematical formulation to set the context for non-specialist developers wanting to use your package.
