### 2022-10-02 #4736 @jbytecode
@jmejia8 - is the issue about the community guidelines okay?

@HaoZeke - could you please update your status?

Thank you in advance.
