### 2022-10-03 JuliaSmoothOptimizers/PDENLPModels.jl#80 @jmejia8
Minor improvements in documentation
I am working on my review for JOSS: https://github.com/openjournals/joss-reviews/issues/4736

The package seems well-coded and structured. However, some minor improvements can be made in the documentation:

- Maybe this table of contents should be inserted in a different section https://github.com/JuliaSmoothOptimizers/PDENLPModels.jl/blob/389eed190bd64958ac6dadeec15a57462feaf422/docs/src/index.md?plain=1#L26
- It might be helpful to provide some bibliographical references for these examples:

https://github.com/JuliaSmoothOptimizers/PDENLPModels.jl/blob/389eed190bd64958ac6dadeec15a57462feaf422/docs/src/poisson-boltzman.md?plain=1#L2

https://github.com/JuliaSmoothOptimizers/PDENLPModels.jl/blob/389eed190bd64958ac6dadeec15a57462feaf422/docs/src/tore.md?plain=1#L2

- Please, check the syntax highlighting in the examples mentioned above.
