### 2022-10-10 dcef2e3
add book ref in intro
