+++ paper/paper.bib
@@ -498,3 +498,11 @@ year = {2021}
   year={2021},
   url={https://github.com/JuliaTopOpt/TopOpt.jl}
 }
+
+@book{hinze2008optimization,
+  title={Optimization with {PDE} constraints},
+  author={Hinze, Michael and Pinnau, Ren{\'e} and Ulbrich, Michael and Ulbrich, Stefan},
+  volume={23},
+  year={2008},
+  publisher={Springer Science \& Business Media}
+}
