### 2022-10-10 b538f32
add refs for brachistochrone problem
