### 2022-10-10 JuliaSmoothOptimizers/PDENLPModels.jl#80 @jmejia8
Solved in https://github.com/JuliaSmoothOptimizers/PDENLPModels.jl/pull/84
@tmigot, thank you!
