### 2022-10-13 2351b3b
add `jac_op!` function
