### 2022-10-20 a431632
Add doi FinEtools (#86)
