+++ paper/paper.bib
@@ -331,7 +331,8 @@ month = {12},
 title = {{FinEtools.jl}},
 url = {https://github.com/PetrKryslUCSD/FinEtools.jl},
 version = {5.3.1},
-year = {2021}
+year = {2021},
+doi = {10.5281/zenodo.7187507}
 }
 
 @misc{JuliaFEM,
