### 2022-10-20 #4736 @tmigot
Dear @jbytecode @jmejia8 @HaoZeke
I re-generated the paper after https://github.com/JuliaSmoothOptimizers/PDENLPModels.jl/pull/86 just to add one of the reference with its new DOI.
