### 2022-11-06 JuliaSmoothOptimizers/PDENLPModels.jl#87 @HaoZeke
JOSS: Review comments [Software Paper]
As written, I'm afraid I do not see any reason that a thin compatibility layer to the already published Gridap should go forward. Please expand the paper with design and implementation considerations.
