### 2022-11-06 #4736 @HaoZeke
I've started (and mostly finished) my review. Many apologies for the delay. It is however unclear to me why a helper wrapper library to the already published Gridap be considered at this stage. I have requested clarifications.
