+++ paper/paper.md
@@ -55,8 +55,8 @@ Additionally, \texttt{Gridap.jl} has an expressive API allowing to model complex
 
 However, the above packages are designed for sets of PDEs and not for optimization, so that only the derivatives of the PDE with respect to $y$ can be evaluated. 
 In addition, inequalities are not supported.
-\texttt{PDENLPModels.jl} extends Gridap's differentiation facilities to also obtain derivatives useful for optimization, i.e., first and second derivatives of the objective and constraint functions with respect to controls and finite-dimensional variables.
-Because we aim to solve nonconvex optimization problems with inequality constraints, it would not be appropriate, or even feasible, to solve a system of PDEs representing the optimality conditions with Gridap.
+\texttt{PDENLPModels.jl} extends \texttt{Gridap.jl}'s differentiation facilities to also obtain derivatives useful for optimization, i.e., first and second derivatives of the objective and constraint functions with respect to controls and finite-dimensional variables.
+Because we aim to solve nonconvex optimization problems with inequality constraints, it would not be appropriate, or even feasible, to solve a system of PDEs representing the optimality conditions with \texttt{Gridap.jl}.
 
 \texttt{PDENLPModels.jl} exports the `GridapPDENLPModel` type, which uses \texttt{Gridap.jl} for the discretization of the functional spaces by finite-elements. The resulting model is an instance of an `AbstractNLPModel`, as defined in \texttt{NLPModels.jl} [@orban-siqueira-nlpmodels-2020], and provides access to objective and constraint function values, to their first and second derivatives, and to any information that a solver might request from a model. As such, \texttt{PDENLPModels.jl} offers an interface between generic PDE-constrained optimization problems and cutting-edge optimization solvers such as \texttt{Artelys Knitro} [@byrd2006k] via \texttt{NLPModelsKnitro.jl} [@orban-siqueira-nlpmodelsknitro-2020], \texttt{Ipopt} [@wachter2006implementation] via \texttt{NLPModelsIpopt.jl} [@orban-siqueira-nlpmodelsipopt-2020], \texttt{DCISolver.jl} [@migot2022dcisolver], \texttt{Percival.jl} [@percival-jl], and any solver accepting an `AbstractNLPModel` as input, see JuliaSmoothOptimizers (JSO) [@jso].
 
