### 2022-11-10 JuliaSmoothOptimizers/PDENLPModels.jl#87 @tmigot
Review for JOSS: https://github.com/openjournals/joss-reviews/issues/4736 Dear @HaoZeke @jbytecode

We do not believe that PDENLPModels qualifies as a thin wrapper of Gridap. Gridap focuses on partial differential equations (only), so that only the derivative of the PDE w.r.t to its solution is available, while PDENLPModels deals with optimization problems (where the PDE is a constraint only) and gives access to the first and second derivatives w.r.t. to the solution of the PDE, the control function and the finite-dimensional variable.

In our opinion, the package respects JOSS's Substantial scholarly effort policy https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort .
The package is feature-complete, the source folder of the packages currently contains 2500 lines of code, designed for maintainable extensions (as we plan to add alternative finite-elements approaches), and offers a significant contribution to the available open-source software as stated in the Statement of need section in the paper.

To make, the contribution of the package clearer w.r.t to Gridap.jl we suggest the following modification: #88
