### 2022-11-10 #4736 @tmigot
@jbytecode @HaoZeke I replied in the issue directly https://github.com/JuliaSmoothOptimizers/PDENLPModels.jl/issues/87
