### 2022-11-10 JuliaSmoothOptimizers/PDENLPModels.jl#87 @jbytecode
Due to the JOSS submission: https://github.com/openjournals/joss-reviews/issues/4736
