### 2022-11-10 #4736 @jbytecode
I think our reviewer has requested clarifications and @tmigot explained why their package is not a thin wrapper and satisfies the JOSS requirements.

@HaoZeke - are those explanations are okay?
