### 2022-11-11 JuliaSmoothOptimizers/PDENLPModels.jl#87 @HaoZeke
I'm afraid I am still a little concerned. In part this is because the [NLPModels themselves](https://github.com/JuliaSmoothOptimizers/NLPModels.jl/blob/main/CITATION.bib), along with the [other wrappers implementing this API](https://juliasmoothoptimizers.github.io/NLPModels.jl/dev/models/) them do not seem to be published. Additionally, LOC is a poor metric, and there is no need (IMO) to nitpick about the code quality in particular (how many lines are code, etc.). It would help to have more than two examples, including an example of extensibility as mentioned.

There are still unfortunately several reasons why then, I'm not sure this is not a "thin" wrapper of existing API implementations, none of which have found the impetus perhaps to be shepherded through JOSS.
