### 2022-11-11 #4736 @HaoZeke
Not yet unfortunately, I have pointed to more concrete arguments about the codebase. I am also happy to hear inputs from you, @jbytecode about the applicability and @jmejia8 at this point.
