### 2022-11-13 #4736 @jmejia8
@tmigot - Please help me by answering the following questions related to @HaoZeke's important comments about the codebase.

1. Which package is handling the main aspects of the modeling? NLPModels, Gridap or yours?
2. Would you say that your package is based on NLPModels or extending NLPModels?
3. When a `GridapPDENLPModel` is created and an optimizer has to solve the modeled problem. Will the optimizer use your API to evaluate the objective function and constraints or will the NLPModels and Gridap APIs be used?
