### 2022-11-18 #4736 @tmigot
The role of NLPModels.jl is to define an API that users and solvers can rely on. It is the role of other packages to implement facilities that create models compliant with the NLPModels API. There are several examples of this in JSO: AMPLNLReader allows users to connect the AMPL modeling language with NLPModels, CUTEst does the same for the SIF language, NLPModelsJuMP does the same with JuMP models, etc. In those three examples, there exist underlying modeling tools (in Julia or not). PDENLPModels is different in that there is no existing generic interface for optimization with PDEs. Instead, PDENLPModels interacts with Gridap to evaluate functionals and differential operators based on a discretization. Gridap in itself does not let users model optimization problems; only systems of PDEs. PDENLPModels provides all the extra facilities for users and solvers to interact with a PDE-constrained optimization problem as they would with a JuMP model, an AMPL model, or any other model that complies with the NLPModels API.

To answer your three questions:

- Both PDENLPModels and Gridap handle the modeling. PDENLPModels interacts with Gridap to evaluate functional/differential objects and their derivatives. However, in addition, PDENLPModels extends the differentiation capabilities of Gridap to address the requirements of optimization solvers, and allows finite-dimensional variables in the problem.
- NLPModels only states the API that we conform to. It's the "syntax" of models, so to speak.
- A solver only needs to worry about the NLPModels API and need never know where a model comes from; all the details are handled inside the model itself. In this instance, PDENLPModels encapsulates the details and interacts with Gridap when necessary. The solver never needs to know about Gridap. Indeed the same solver can be used to solve an AmplModel, a CUTEstModel, a MathProgNLPModel, etc.

We would be happy to add the above clarifications to the paper if you deem it necessary.
