### 2022-11-18 #4736 @jmejia8
Regarding the responses from @tmigot and my review of the source code, I've found significant contributions to the mathematical programming community to solve such complicated problems. Finding a suitable way to model nonlinear programming problems with PDE constraints is, for me, a worthwhile research software.

@tmigot - The clarification you made on the PDENLPModels API could be useful to have in the package documentation.
