### 2022-11-29 #4736 @HaoZeke
@tmigot as suggested by @jmejia8 as well, could you update the package documentation and paper before we proceed?
