### 2022-12-01 516d5d2
add more details on NLPModel API [temp]
