### 2022-12-08 a2dadcd
Exp2 (#89)
* add more details on NLPModel API
