+++ paper/paper.md
@@ -58,7 +58,10 @@ In addition, inequalities are not supported.
 \texttt{PDENLPModels.jl} extends \texttt{Gridap.jl}'s differentiation facilities to also obtain derivatives useful for optimization, i.e., first and second derivatives of the objective and constraint functions with respect to controls and finite-dimensional variables.
 Because we aim to solve nonconvex optimization problems with inequality constraints, it would not be appropriate, or even feasible, to solve a system of PDEs representing the optimality conditions with \texttt{Gridap.jl}.
 
-\texttt{PDENLPModels.jl} exports the `GridapPDENLPModel` type, which uses \texttt{Gridap.jl} for the discretization of the functional spaces by finite-elements. The resulting model is an instance of an `AbstractNLPModel`, as defined in \texttt{NLPModels.jl} [@orban-siqueira-nlpmodels-2020], and provides access to objective and constraint function values, to their first and second derivatives, and to any information that a solver might request from a model. As such, \texttt{PDENLPModels.jl} offers an interface between generic PDE-constrained optimization problems and cutting-edge optimization solvers such as \texttt{Artelys Knitro} [@byrd2006k] via \texttt{NLPModelsKnitro.jl} [@orban-siqueira-nlpmodelsknitro-2020], \texttt{Ipopt} [@wachter2006implementation] via \texttt{NLPModelsIpopt.jl} [@orban-siqueira-nlpmodelsipopt-2020], \texttt{DCISolver.jl} [@migot2022dcisolver], \texttt{Percival.jl} [@percival-jl], and any solver accepting an `AbstractNLPModel` as input, see JuliaSmoothOptimizers (JSO) [@jso].
+\texttt{PDENLPModels.jl} exports the `GridapPDENLPModel` type, which uses \texttt{Gridap.jl} for the discretization of the functional spaces by finite-elements. The resulting model is an instance of an `AbstractNLPModel`, as defined in \texttt{NLPModels.jl} [@orban-siqueira-nlpmodels-2020], and provides access to objective and constraint function values, to their first and second derivatives, and to any information that a solver might request from a model. 
+
+The role of \texttt{NLPModels.jl} is to define an API that users and solvers can rely on. It is the role of other packages to implement facilities that create models compliant with the NLPModels API. There are several examples of this in JuliaSmoothOptimizers organization: \texttt{AmplNLReader.jl} [@orban-siqueira-amplnlreader-2020] allows users to connect the AMPL modeling language with NLPModels, \texttt{CUTEst.jl} [@orban-siqueira-cutest-2020] does the same for the SIF language, \texttt{NLPModelsJuMP.jl} [@montoison-orban-siquiera-nlpmodelsjump-2020] does the same with JuMP models, etc. In those three examples, there exist underlying modeling tools (in Julia or not). \texttt{PDENLPModels.jl} is different in that there is no existing generic interface for optimization with PDEs. Instead, \texttt{PDENLPModels.jl} interacts with \texttt{Gridap.jl} to evaluate functionals and differential operators based on a discretization. \texttt{Gridap.jl} in itself does not let users model optimization problems; only systems of PDEs. \texttt{PDENLPModels.jl} provides all the extra facilities for users and solvers to interact with a PDE-constrained optimization problem as they would with a JuMP model, an AMPL model, or any other model that complies with the NLPModels API.
[…]
