### 2022-12-09 #4736 @tmigot
Dear @jbytecode , @jmejia8 , @HaoZeke,

Sorry for the late update.

I opened two PRs adding further comments on the NLPModels framework following our discussion:
- for the paper: https://github.com/JuliaSmoothOptimizers/PDENLPModels.jl/pull/88
- for the documentation: https://github.com/JuliaSmoothOptimizers/PDENLPModels.jl/pull/90

If necessary, for instance, to generate the pdf in the JOSS thread, I can merge any of them.
