### 2022-12-09 309d4b3
Expand contribution (#88)
* expand on the contribution

* add more details on NLPModel API
