+++ paper/paper.md
@@ -48,9 +48,20 @@ After discretization of the domain $\Omega$, the integral, and the derivatives, 
 \end{equation*}
 where $l, u \in \mathbb{R}^{N_y + N_u + N_\theta}$, $f:\mathbb{R}^{N_y} \times \mathbb{R}^{N_u} \times \mathbb{R}^{N_\theta} \rightarrow \mathbb{R}$ and $c:\mathbb{R}^{N_y} \times \mathbb{R}^{N_u} \times \mathbb{R}^{N_\theta} \rightarrow \mathbb{R}^{N_y}$.
 
-The first difficulty in modeling such a challenging problem is to access a discretization of the domain and have the possibility to evaluate derivatives of $f$ and $c$. Fortunately, several packages allow the user to define the domain, meshes, function spaces, and finite-element families to approximate unknowns and model functionals and sets of PDEs in a weak form. The main ones are \texttt{FEniCS.jl}, a wrapper for the FEniCS library [@logg2012automated], \texttt{Ferrite.jl} [@Carlsson_Ferrite_jl_2021], \texttt{FinEtools.jl} [@Krysl_FinEtools_jl_2021], \texttt{JuliaFEM.jl} [@JuliaFEM] [@JuliaFEM-paper], and \texttt{Gridap.jl} [@badia2020gridap] [@verdugo2022software]. In \texttt{PDENLPModels.jl}, we focus on the latter as it is exclusively written in Julia and supports a variety of discretizations and meshing possibilities.  Additionally, \texttt{Gridap.jl} has an expressive API allowing to model complex PDEs with few lines of code, and to write the underlying weak form with a syntax almost one-to-one with mathematical notation.
+The two main challenges in modeling such a problem are to be able to (i) discretize the domain and generate corresponding discretizations of the objective and constraints, and (ii) evaluate derivatives of $f$ and $c$ with respect to all variables.
+Several packages allow the user to define the domain, meshes, function spaces, and finite-element families to approximate unknowns, and to model functionals and sets of PDEs in a weak form using powerful high-level notation. The main ones are \texttt{FEniCS.jl}, a wrapper for the FEniCS library [@logg2012automated], \texttt{Ferrite.jl} [@Carlsson_Ferrite_jl_2021], \texttt{FinEtools.jl} [@Krysl_FinEtools_jl_2021], \texttt{JuliaFEM.jl} [@JuliaFEM] [@JuliaFEM-paper], and \texttt{Gridap.jl} [@badia2020gridap] [@verdugo2022software].
+In \texttt{PDENLPModels.jl}, we focus on the latter as it is exclusively written in Julia and supports a variety of discretizations and meshing possibilities.
+Additionally, \texttt{Gridap.jl} has an expressive API allowing to model complex PDEs with few lines of code, and to write the underlying weak form with a syntax that is almost one-to-one with mathematical notation.
 
[…]
