### 2022-12-13 #4736 @HaoZeke
I'm OK with the current formulation of the paper and code. I'm happy to recommend this move forward and thank the author for taking time to iterate on the resolutions.
