### 2022-12-13 #4736 @jbytecode
@HaoZeke - Thank you for the response. The item

> Substantial scholarly effort: Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)

is still unselected. Is there still some work waiting for resolving this issue or you've forgotten to check it on?

- Please check this on if everything is okay with this item
- Please redirect the author to fill the gaps in paper and code if still exist.

Thank you in advance.
