### 2022-12-16 #4736 @jbytecode
@tmigot - The paper and bibliography look nice to me for now on. A small capitalization issue should be corrected in bibtex. I have just sent a pull request. Could you please apply it? (https://github.com/JuliaSmoothOptimizers/PDENLPModels.jl/pull/91)
