### 2022-12-16 #4736 @tmigot
Thanks @jbytecode I just merged it.
