### 2022-12-16 #4736 @jbytecode
@tmigot - an additional problem with the README.md file: The mathematical expression under the sentence

`We consider optimization problems of the form:`

overflows and does not fit the width of the view. You may try multiline equations or latex environment `aligned` to fix it.

Thank you in advance.
