### 2022-12-16 f9aab79
latexify math expression in README.md
