+++ paper/paper.bib
@@ -23,6 +23,33 @@
   doi={10.1007/s10107-004-0559-y}
 }
 
+@software{orban-siqueira-cutest-2020,
+  author = {D. Orban and A. S. Siqueira and {contributors}},
+  title = {{CUTEst.jl}: {J}ulia's {CUTEst} interface},
+  month = {October},
+  url = {https://github.com/JuliaSmoothOptimizers/CUTEst.jl},
+  year = {2020},
+  DOI = {10.5281/zenodo.1188851},
+}
+
+@software{orban-siqueira-amplnlreader-2020,
+  author = {D. Orban and A. S. Siqueira and {contributors}},
+  title = {{AmplNLReader.jl}: A {J}ulia interface to {AMPL}},
+  month = {July},
+  url = {https://github.com/JuliaSmoothOptimizers/AmplNLReader.jl},
+  year = {2020},
+  DOI = {10.5281/zenodo.3700941},
+}
+
+@software{montoison-orban-siquiera-nlpmodelsjump-2020,
+  author = {A. Montoison and D. Orban and A. S. Siqueira},
+  title = {{NLPModelsJuMP.jl}: Conversion from {JuMP} models to {NLPModels}},
+  month = {August},
+  url = {https://github.com/JuliaSmoothOptimizers/NLPModelsJuMP.jl},
+  year = {2020},
+  DOI = {10.5281/zenodo.2574162},
+}
+
 @software{percival-jl,
   author       = {Egmara Antunes dos Santos and Abel Soares Siqueira},
   title        = {{Percival.jl}: an augmented Lagrangian method},
