+++ paper/paper.bib
@@ -300,7 +300,7 @@ doi="10.1007/0-387-30065-1_4",
 }
 
 @article{ruthotto2017jinv,
-title={j{I}nv--a flexible {J}ulia package for {PDE} parameter estimation},
+title={{jInv} -- a flexible {J}ulia package for {PDE} parameter estimation},
 author={Ruthotto, Lars and Treister, Eran and Haber, Eldad},
 journal={SIAM Journal on Scientific Computing},
 volume={39},
