+++ Project.toml
@@ -1,7 +1,7 @@
 name = "PDENLPModels"
 uuid = "80da258d-c172-4540-9646-64f172512520"
 authors = ["tmigot <tangi.migot@gmail.com>"]
-version = "0.3.2"
+version = "0.3.3"
 
 [deps]
 FastClosures = "9aa1b823-49e4-5ca5-8b0f-3971ec8bab6a"
