### 2022-12-16 #4736 @jbytecode
@tmigot - we have one more capitalization issue in the bibtex file. please merge it. you don't need to re-generate a zenodo archive after all. thank you in advance and I am sorry for the inconvenience.
