### 2022-12-16 2333875
Capitalization in bibtex (#94)
* capitalization in bibtex
