+++ paper/paper.bib
@@ -366,7 +366,7 @@ doi = {10.5281/zenodo.7187507}
   doi = {10.5281/zenodo.1410189},
   url = {https://github.com/JuliaFEM/JuliaFEM.jl},
   author = {Aho,  Jukka and Frondelius,  Tero and {Ovainola} and {Arilaakk} and Kelman,  Tony and Stoian,  Vlad and Badger,  The Gitter and Rapo,  Marja},
-  title = {Juliafem/Juliafem.Jl: Julia V1 Compatible Release},
+  title = {{JuliaFEM}/{JuliaFEM.jl}: {Julia} V1 Compatible Release},
   publisher = {Zenodo},
   year = {2018},
 }
