+++ paper/paper.md
@@ -51,14 +51,14 @@ where $l, u \in \mathbb{R}^{N_y + N_u + N_\theta}$, $f:\mathbb{R}^{N_y} \times \
 The two main challenges in modeling such a problem are to be able to (i) discretize the domain and generate corresponding discretizations of the objective and constraints, and (ii) evaluate derivatives of $f$ and $c$ with respect to all variables.
 Several packages allow the user to define the domain, meshes, function spaces, and finite-element families to approximate unknowns, and to model functionals and sets of PDEs in a weak form using powerful high-level notation. The main ones are \texttt{FEniCS.jl}, a wrapper for the FEniCS library [@logg2012automated], \texttt{Ferrite.jl} [@Carlsson_Ferrite_jl_2021], \texttt{FinEtools.jl} [@Krysl_FinEtools_jl_2021], \texttt{JuliaFEM.jl} [@JuliaFEM] [@JuliaFEM-paper], and \texttt{Gridap.jl} [@badia2020gridap] [@verdugo2022software].
 In \texttt{PDENLPModels.jl}, we focus on the latter as it is exclusively written in Julia and supports a variety of discretizations and meshing possibilities.
-Additionally, \texttt{Gridap.jl} has an expressive API allowing to model complex PDEs with few lines of code, and to write the underlying weak form with a syntax that is almost one-to-one with mathematical notation.
+Additionally, \texttt{Gridap.jl} has an expressive API allowing one to model complex PDEs with few lines of code, and to write the underlying weak form with a syntax that is almost one-to-one with mathematical notation.
 
 However, the above packages are designed for sets of PDEs and not for optimization, so that only the derivatives of the PDE with respect to $y$ can be evaluated. 
 In addition, inequalities are not supported.
 \texttt{PDENLPModels.jl} extends \texttt{Gridap.jl}'s differentiation facilities to also obtain derivatives useful for optimization, i.e., first and second derivatives of the objective and constraint functions with respect to controls and finite-dimensional variables.
 Because we aim to solve nonconvex optimization problems with inequality constraints, it would not be appropriate, or even feasible, to solve a system of PDEs representing the optimality conditions with \texttt{Gridap.jl}.
 
-\texttt{PDENLPModels.jl} exports the `GridapPDENLPModel` type, which uses \texttt{Gridap.jl} for the discretization of the functional spaces by finite-elements. The resulting model is an instance of an `AbstractNLPModel`, as defined in \texttt{NLPModels.jl} [@orban-siqueira-nlpmodels-2020], and provides access to objective and constraint function values, to their first and second derivatives, and to any information that a solver might request from a model. 
+\texttt{PDENLPModels.jl} exports the `GridapPDENLPModel` type, which uses \texttt{Gridap.jl} for the discretization of the functional spaces by finite-elements. The resulting model is an instance of an `AbstractNLPModel`, as defined in \texttt{NLPModels.jl} [@orban-siqueira-nlpmodels-2020], that provides access to objective and constraint function values, to their first and second derivatives, and to any information that a solver might request from a model. 
 
[…]
