+++ paper/paper.bib
@@ -51,8 +51,8 @@
 }
 
 @software{percival-jl,
-  author       = {Egmara Antunes dos Santos and Abel Soares Siqueira},
-  title        = {{Percival.jl}: an augmented Lagrangian method},
+  author       = {dos Santos, Egmara Antunes and Abel Soares Siqueira},
+  title        = {{Percival.jl}: an augmented {L}agrangian method},
   month        = jul,
   year         = 2020,
   url          = {https://github.com/JuliaSmoothOptimizers/Percival.jl},
