+++ paper/paper.bib
@@ -51,7 +51,7 @@
 }
 
 @software{percival-jl,
-  author       = {dos Santos, Egmara Antunes and Abel Soares Siqueira},
+  author       = {Egmara Antunes {dos Santos} and Abel Soares Siqueira},
   title        = {{Percival.jl}: an augmented {L}agrangian method},
   month        = jul,
   year         = 2020,
