### 2022-12-16 #4736 @danielskatz
@tmigot - sorry, my first fix for the surname in the bibtex didn't work (though I'm not sure why not) - can we try https://github.com/JuliaSmoothOptimizers/PDENLPModels.jl/pull/96 next?
